requirements.txt
pyrogram==2.0.106
python-telegram-bot==20.7
pdf2image==1.16.3
pymupdf==1.22.3
Pillow==10.2.0
